+++ src/main/java/racingcar/controller/ConsoleController.java
@@ -1,7 +1,7 @@
 package racingcar.controller;
 
 import racingcar.domain.Car;
-import racingcar.domain.ConsoleService;
+import racingcar.service.ConsoleService;
 import racingcar.domain.RandomMoveChance;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
