+++ src/main/java/racingcar/controller/WebController.java
@@ -3,7 +3,7 @@ package racingcar.controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import racingcar.domain.WebService;
+import racingcar.service.WebService;
 import racingcar.dto.RequestDto;
 import racingcar.dto.ResponseDto;
